feat: 엔티티 생성

+++ src/main/java/induckmoms/server/common/config/AppConfig.java
@@ -41,7 +41,7 @@ public class AppConfig {
   }
 
   @Bean
-  public AuditorAware<Integer> auditorAware() {
+  public AuditorAware<Long> auditorAware() {
     return new ApplicationAuditAware();
   }
